backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
bleach==6.2.0
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.3
django-cors-headers==4.9.0
django-unfold==0.83.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
idna==3.11
kombu==5.6.2
numpy==2.0.2
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
gunicorn==22.0.0
redis==7.0.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
sslcommerz-lib==1.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0

beautifulsoup4==4.12.3
lxml==5.3.0
whitenoise==6.9.0
django-redis==5.4.0
